+++ lib/src/optional_arg_validator.dart
@@ -1,0 +1,10 @@
+import 'package:dartle/dartle.dart' show ArgsCount;
+
+class OptionalArgValidator extends ArgsCount {
+  final String help;
+
+  const OptionalArgValidator(this.help) : super.range(min: 0, max: 1);
+
+  @override
+  String helpMessage() => help;
+}